Add else to if statements for fluidic compressor canister stacking if statements
Hi, I'm new to modding here.
Just thought I would note this issue with the fluidic compressor
crashing when finishing filling or emptying a canister.
I added else's to the second if statements to prevent trying to
reference the input stack when it is set as NULL by the previous if
statement.